Check for small payload_len when parsing extend cells.
Without this code, if V1 relay cell format were in use, and the relay message
length were set to 0 or 1, then an EXTEND cell could be read beyond the end of
the relay cell payload.  This could extend beyond the underlying cell body if
the V1 relay cell format had been negotiated.  This would typically lead either
to a crash or to a rejected circuit.

Closes bug 41180; bugfix on 0.4.9.3-alpha, when we made the made the maximum
size of a relay payload variable.

0.4.8.x and earlier can similarly mis-handle low payload values, but the
bug there cannot be used to read uninitialized data.